+++ example/main.ts
@@ -1,1 +1,5 @@
-/* eslint-disable */
+/* eslint-disable capitalized-comments, line-comment-position */
+import {Component, OnInit} from '@angular/core';
+import {Cookies} from '@cedx/ngx-cookies';
+
+TODO !!!!!!!